Switched to bootstrap 3
Hey all sorry for the huge delay - I was having a issue with some CSS that was messing with images. I dont give a shit about the design rn so hack away. Mostly just wanted to get the bootstrap version pushed.

I'm using LESS in this project. If you check out the CSS file it will be pretty undreadable, so make sure to make changes in the LESS file. Any changes to the CSS will just be overwritten when LESS compiles.

Btw you may notice a strange comment at the top of the LESS file. I'm using Brackets.io for this project, and that tells the LESS compiler where to put the compiled CSS.